OSIS-1953 move filter in form
